sarzak.toml
Funny, I called the branch cargo-dot-toml. It's hard to break a habit.
sarzak new adds the domain to sarzak.toml, although not in a very
pretty manner. And I can't get it working as advertised.

sarzak gen does all sorts of neat stuff. But primarily it combines info
from the config with cl params and does it's job. No longer does it
look over files in models, but instead over domains in config.

Other stuff, like checking if a module directory exists. It's probably
buggy, and they will surface with use. For now it's time to put a fork
in it.